Issue #6: updated gridview and home_screen
