+++ lib/main.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 //import 'package:recipe_organizer_frontend/app.dart';
-import 'package:recipe_organizer_frontend/screens/navbar_android.dart';
+import 'package:recipe_organizer_frontend/screens/home_screen_android.dart';
 import 'package:recipe_organizer_frontend/screens/recipe_detail_screen.dart';
 
 void main() {
@@ -21,11 +21,13 @@ class MyApp extends StatelessWidget {
         ),
       ),
       //home: kIsWeb ? MyHomePageWeb() : MyHomePageApp(),
-      home: Detailspage(
+      home: MyHomePageApp(),
+      /*home: Detailspage(
         image:"https://images.unsplash.com/photo-1586190848861-99aa4a171e90?q=80&w=2080&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D", 
         name: "Cheeseburger", 
         username: "Moser",
         userimage: "https://images.unsplash.com/photo-1586190848861-99aa4a171e90?q=80&w=2080&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D"),
+    */
     );
   }
 }
