+++ lib/screens/login_page.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:recipe_organizer_frontend/screens/navbar_android.dart';
 import '../utils/api.dart';
 
 class LoginPage extends StatefulWidget {
