+++ lib/screens/navbar_android.dart
@@ -1,93 +1,48 @@
-import 'package:flutter/cupertino.dart';
-import 'package:flutter/foundation.dart';
-import 'package:google_nav_bar/google_nav_bar.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/cupertino.dart';
+import 'package:google_nav_bar/google_nav_bar.dart';
 import 'login_page.dart';
-import 'search_bar.dart';
-import 'gridview.dart';
+import '../colors.dart';
 
-class MyHomePageApp extends StatefulWidget{
-  const MyHomePageApp({Key? key}): super(key: key);
+class NavBarScreen extends StatelessWidget {
+  const NavBarScreen({super.key});
 
   @override
-  State<MyHomePageApp> createState() => _MyHomePageAppState();
-}
-
-
-
-class _MyHomePageAppState extends State<MyHomePageApp>{
-  @override
-  Widget build(BuildContext context){
-    return Scaffold(
-      
-       body: SafeArea(
-         child: SingleChildScrollView(
-           child: Column(
-             children: [
-                   Container(
-                    constraints: BoxConstraints(maxHeight: 90.0),
-                     child: Padding(
-                       padding: EdgeInsets.all(8.0),
-                       child: SearchBarApp(),
-                     ),
-                   ),
-                   //SizedBox(height: 1000,),
-                   Padding(
-                     padding: EdgeInsets.all(8.0),
-                     child: GridB(),
-                   ),
-             ],
-           ),
-         ),
-       ),
-      bottomNavigationBar: Container(
-        color: const Color.fromRGBO(121,220,241,1),
-        child: Padding(
-          padding: const EdgeInsets.all(15.0),
-          child: GNav(
-            backgroundColor: const Color.fromRGBO(121,220,241,1),
-            color: Colors.white,
-            activeColor: Colors.white,
-            tabBackgroundColor: Colors.grey.shade800,
-            padding: const EdgeInsets.all(16),
-            onTabChange: (index){
-              if(index == 3){
-                Navigator.push(
-                  context,
-                  MaterialPageRoute(
-                    builder: (context) => const LoginPage(title: 'Recipe Organizer: Login Page')
-                    ),
-          );}
-            },
-          
-            gap: 8,
-            tabs: const [
-              GButton(icon: Icons.home, text: 'Home',),
-              GButton(icon: Icons.favorite, text: 'Liked Recipes'),
-              GButton(icon: Icons.schedule, text: 'Meal Plan'),
-              GButton(icon: Icons.person, text: 'Profile')
-              ],
+  Widget build(BuildContext context) {
+    return
+        Container(
+          color: primary,
+          child: Padding(
+            padding: const EdgeInsets.all(15.0),
+            child: GNav(
+              backgroundColor: primary,
+              color: Colors.white,
+              activeColor: Colors.white,
+              tabBackgroundColor: Colors.grey.shade800,
+              padding: const EdgeInsets.all(16),
+              onTabChange: (index){
+                if(index == 3){
+                  Navigator.push(
+                    context,
+                    MaterialPageRoute(
+                      builder: (context) => const LoginPage(title: 'Recipe Organizer: Login Page')
+                      ),
+                    );
+                }
+                if(index == 0){
+                  Navigator.popUntil(context, (route) => route.isFirst);
+                }
+                },
+            
+              gap: 8,
+              tabs: const [
+                GButton(icon: Icons.home, text: 'Home',),
+                GButton(icon: Icons.favorite, text: 'Liked Recipes'),
+                GButton(icon: Icons.schedule, text: 'Meal Plan'),
+                GButton(icon: Icons.person, text: 'Profile')
+                ],
+            ),
           ),
-        ),
-      ),
-      floatingActionButton: FloatingActionButton(
-        onPressed: () {},
-        backgroundColor: const Color.fromRGBO(121,220,241,1),
-        child: const Icon(Icons.add),
-      
-      ),
-      appBar: AppBar(
-        title: const Text("Recipe Organizer"),
-        titleTextStyle: const TextStyle(color: Colors.white, fontSize: 24.0),
-        backgroundColor: Colors.black,
-        actions: <Widget>[
-          IconButton(
-            onPressed: () {}, 
-            icon: const Icon(Icons.search),
-            color: Colors.white,)
-        ],
-        ),
-       
-    );
+        );
   }
 }
